• Overhaul run_xcode_target.rb. It's now a library, providing a simple abstraction for Xcode projects. If you execute the file directly, it will continue to function as it has previously.
This means we can nicely display errors and status in HTML for 'Build & Run'.

Even better would be directly executing the application, allowing console output to be printed in the build output window. But this seems to be impossible without parsing build configurations. So not yet.

(Tools should probably continue to run in an external Terminal window, they're more likely to require stdin.)

Eventually, there should be UI for choosing an executable to run in the case of multitarget projects. Right now, we just bail out, no change.

git-svn-id: http://svn.textmate.org/trunk/Bundles/Xcode.tmbundle@6150 dfb7d73b-c2ec-0310-8fea-fb051d288c6d